Declare get_db function in database.py
